Updated readme.md: Avahi patch, Mopria/Wi-Fi Direct printer support
The patch for Avahi, suggested in readme.md, was only making Avahi
advertise local-only services, like IPP-over-USB printers getting
advertised at all, but they got advertised with the network host name
of the machine, not with "localhost", making auto-generated print
queues not working. The fix for this is found now and the patch
appropriately updated.

ippusbxd does not only support Apple AirPrint and IPP Everywhere
ptinters but also the other driverless printing standards, Mopria and
Wi-Fi Direct. This is now also mentioned in readme.md, also the fact
that the "-meverywhere" argument of CUPS' "lpadmin" command does not
support PCLm.