Fix automatic search of an IPP-over-USB printer
When starting "ippusbxd" without specification of the printer (like
Vendor/Product ID, Bus/Device, and/or serial number), it is supposed
to search for the first IPP-over-USB printer under the USB
devices. This did not work. The search always stopped at the first
unsuitable device. This change solves the problem.