Use URF info in the USB device ID for the DNS-SD record
Some printers include the URF info string ("urf-supported" IPP
attribute, tells about printer capabilities which can be used in Apple
Raster mode) in their USB device ID as the "URF" field. If so, make
use of it for the DNS-SD record to be broadcasted. Once, include the
string itself as "URF" TXT record field, and second, set the boolean
TXT record fields "Duplex", "Color", and "Copies" according to the URF
info string.