Changing unregister_service_thread to deallocate removed thread params
The service_thread_param parameter for a communication thread was being
unallocated before calling to the thread cleanup handler. This sometimes
would result in an error as the entry in the service_threads list would
become invalid causing problems when attempting to unregister threads.

To fix this problem, the thread cleanup handler
(unregister_service_thread()) has been changed to also take care of
freeing the memory associated with the service_thread_param which was
removed.