Reducing logging code and making fixes to web scanning
This patch changes the tcp_packet_get() function so that if an attempted
read from the tcp socket were to time out while the printer is still
transferring data then the connection will remain open.

This should fix any issues with high dpi scanning, as in earlier
versions the socket was being closed prematurely while the printer was
still streaming data over the socket from the scan.

This patch also cuts down on some of the unnecessary logging code.