Fixing spurious recv timeout errors and fixing compile warnings
When using the web interface, the web browser will sometimes hold on to
a connection on a tcp socket even though it is not currently
transferring any data. When trying to read from such a connection the
recv system call would fail with either EAGAIN or EWOULDBLOCK indicating
that the attempt to read from the socket had timed out. This patch
changes the attempted reads from the socket to use polling so that a
timeout should only occur in the event of an actual error.